Moving BaseTree into own file under tree module
